+++ Staging/AudioDxe/AudioDxe.c
@@ -32,8 +32,9 @@
 #include <Library/OcBootManagementLib.h>
 #include <Library/OcFlexArrayLib.h>
 
-UINTN	gGpioSetupStageMask   = GPIO_SETUP_STAGE_NONE;
-UINTN	gGpioPinMask          = GPIO_PIN_MASK_AUTO;
+UINTN	  gGpioSetupStageMask = GPIO_SETUP_STAGE_NONE;
+UINTN	  gGpioPinMask        = GPIO_PIN_MASK_AUTO;
+BOOLEAN gRestoreNoSnoop     = FALSE;
 
 /**
   HdaController Driver Binding.
@@ -87,11 +88,13 @@ AudioDxeInit(
 
   Status = OcParseLoadOptions (LoadedImage, &ParsedLoadOptions);
   if (!EFI_ERROR (Status)) {
+    gRestoreNoSnoop = OcHasParsedVar (ParsedLoadOptions, L"--restore-nosnoop", TRUE);
+
     Status = OcParsedVarsGetInt (ParsedLoadOptions, L"--gpio-setup", &gGpioSetupStageMask, TRUE);
     if (Status == EFI_NOT_FOUND && OcHasParsedVar (ParsedLoadOptions, L"--gpio-setup", TRUE)) {
       gGpioSetupStageMask = GPIO_SETUP_STAGE_ALL;
     }
-    DEBUG ((DEBUG_INFO, "HDA: GPIO setup stages 0x%X\n", gGpioSetupStageMask));
+    DEBUG ((DEBUG_INFO, "HDA: GPIO setup stages 0x%X, restore NSNPEN %d\n", gGpioSetupStageMask, gRestoreNoSnoop));
 
     if (gGpioSetupStageMask != GPIO_SETUP_STAGE_NONE) {
       OcParsedVarsGetInt (ParsedLoadOptions, L"--gpio-pins", &gGpioPinMask, TRUE);